feat: #2098 refactor selectors
