+++ packages/marketplace/src/components/pages/app-detail/client/app-detail.tsx
@@ -6,9 +6,9 @@ import { useHistory } from 'react-router'
 import AppUninstallConfirmation from '@/components/pages/app-detail/client/app-uninstall-confirmation'
 import { DesktopIntegrationTypeModel } from '@/actions/desktop-integration-types'
 import { AppDetailDataNotNull } from '@/reducers/client/app-detail'
-import { selectIntegrationTypes } from '@/selector/integration-types'
+import { selectIntegrationTypes } from '@/selector/desktop-integration-types'
 import { useSelector } from 'react-redux'
-import { selectAppDetailData, selectAppDetailLoading } from '@/selector/client-app-detail'
+import { selectAppDetailData, selectAppDetailLoading } from '@/selector/apps'
 import { selectClientId, selectIsAdmin } from '@/selector/auth'
 import { canGoBack } from '@/utils/router-helper'
 import AppContent from './app-content'
@@ -24,7 +24,7 @@ import { BackToAppsSection } from '../common/ui-sections'
 import { AppDetailButtonGroup } from './app-detail-button-group'
 import { useReapitConnect } from '@reapit/connect-session'
 import { reapitConnectBrowserSession } from '@/core/connect-session'
-import { selectDeveloperEditionId } from '@/selector/client'
+import { selectDeveloperEditionId } from '@/selector/auth'
 
 export type ClientAppDetailProps = {}
