+++ packages/marketplace/src/components/pages/app-management/apps-management.tsx
@@ -6,7 +6,7 @@ import { Loader, Info, Pagination, H3 } from '@reapit/elements'
 import ErrorBoundary from '@/components/hocs/error-boundary'
 import routes from '@/constants/routes'
 import AppList from '@/components/ui/app-list'
-import { selectMyApps } from '@/selector/client'
+import { selectMyApps } from '@/selector/apps'
 import { AppSummaryModel } from '@reapit/foundations-ts-definitions'
 import { handleLaunchApp } from '@/utils/launch-app'
 import { selectDeveloperId, selectIsAdmin } from '@/selector/auth'
