+++ packages/marketplace/src/components/ui/app-sidebar/app-sidebar.tsx
@@ -6,7 +6,7 @@ import styles from '@/styles/blocks/app-sidebar.scss?mod'
 import CategoriesList from '@/components/ui/categories-list'
 import { FaSearch } from 'react-icons/fa'
 import { CategoryModel } from '@reapit/foundations-ts-definitions'
-import { selectCategories } from '@/selector/app-categories'
+import { selectCategories } from '@/selector/categories'
 import { addQuery, removeQuery, getParamValueFromPath } from '@/utils/client-url-params'
 import { cleanObject } from '@reapit/utils'
 import { validationSchema } from './validation-schema'
