+++ packages/marketplace/src/sagas/apps/__tests__/installed-apps.ts
@@ -7,7 +7,7 @@ import { Action } from '@/types/core'
 import { cloneableGenerator } from '@redux-saga/testing-utils'
 import { errorThrownServer } from '@/actions/error'
 import errorMessages from '@/constants/error-messages'
-import { selectDeveloperEditionId } from '@/selector/client'
+import { selectDeveloperEditionId } from '@/selector/auth'
 import { fetchAppsList } from '@/services/apps'
 import { INSTALLED_APPS_PERPAGE } from '@/constants/paginator'
 import { selectClientId } from '@/selector/auth'
