+++ packages/marketplace/src/sagas/apps/__tests__/my-apps.ts
@@ -7,7 +7,7 @@ import { Action } from '@/types/core'
 import { cloneableGenerator } from '@redux-saga/testing-utils'
 import { APPS_PER_PAGE } from '@/constants/paginator'
 import errorMessages from '@/constants/error-messages'
-import { selectDeveloperEditionId } from '@/selector/client'
+import { selectDeveloperEditionId } from '@/selector/auth'
 import { errorThrownServer } from '@/actions/error'
 import { fetchAppsList } from '@/services/apps'
 import { reapitConnectBrowserSession } from '@/core/connect-session'
