+++ packages/marketplace/src/sagas/apps/client.ts
@@ -6,8 +6,9 @@ import { errorThrownServer } from '@/actions/error'
 import errorMessages from '@/constants/error-messages'
 import { FEATURED_APPS } from '@/constants/paginator'
 import { Action } from '@/types/core'
-import { selectFeaturedApps, selectDeveloperEditionId } from '@/selector/client'
-import { selectCategories } from '@/selector/app-categories'
+import { selectFeaturedApps } from '@/selector/apps'
+import { selectDeveloperEditionId } from '@/selector/auth'
+import { selectCategories } from '@/selector/categories'
 import { ClientAppSummary, ClientAppSummaryParams } from '@/reducers/client/app-summary'
 import { logger } from '@reapit/utils'
 import { fetchAppsList } from '@/services/apps'
