+++ packages/marketplace/src/sagas/apps/installed-apps.ts
@@ -5,7 +5,7 @@ import { errorThrownServer } from '@/actions/error'
 import errorMessages from '@/constants/error-messages'
 import { Action } from '@/types/core'
 import { INSTALLED_APPS_PERPAGE } from '@/constants/paginator'
-import { selectDeveloperEditionId } from '@/selector/client'
+import { selectDeveloperEditionId } from '@/selector/auth'
 import { logger } from '@reapit/utils'
 import { fetchAppsList } from '@/services/apps'
 import { reapitConnectBrowserSession } from '@/core/connect-session'
