+++ packages/marketplace/src/sagas/installations/__tests__/app-installations.ts
@@ -6,7 +6,7 @@ import { put, takeLatest, all, fork, call } from '@redux-saga/core/effects'
 import { Action } from '@/types/core'
 import { cloneableGenerator } from '@redux-saga/testing-utils'
 import { UninstallParams, InstallParams, appInstallationsSetFormState } from '@/actions/installations'
-import { selectLoggedUserEmail } from '@/selector/client'
+import { selectLoggedUserEmail } from '@/selector/auth'
 import { createInstallation, removeAccessToAppById } from '@/services/installations'
 import { reapitConnectBrowserSession } from '@/core/connect-session'
 import { ReapitConnectSession } from '@reapit/connect-session'
