+++ packages/marketplace/src/sagas/installations/app-installations.ts
@@ -4,7 +4,7 @@ import { Action } from '@/types/core'
 import { errorThrownServer } from '@/actions/error'
 import errorMessages from '@/constants/error-messages'
 import { appInstallationsSetFormState, UninstallParams, InstallParams } from '@/actions/installations'
-import { selectLoggedUserEmail } from '@/selector/client'
+import { selectLoggedUserEmail } from '@/selector/auth'
 import { logger } from '@reapit/utils'
 import { createInstallation, removeAccessToAppById } from '@/services/installations'
 import { reapitConnectBrowserSession } from '@/core/connect-session'
