+++ packages/marketplace/src/selector/apps/__tests__/app.test.ts
@@ -1,15 +1,20 @@
+import { appDetailDataStub, appAuthenticationStub } from '@/sagas/__stubs__/app-detail'
+import { ReduxState } from '@/types/core'
+import appState from '@/reducers/__stubs__/app-state'
+import { featuredAppsDataStub } from '@/sagas/__stubs__/apps'
 import {
   selectAppDetailId,
   selectAppDetailInstallationId,
   selectApp,
   selectAppAuthenticationLoading,
   selectAppAuthenticationCode,
-} from '../app-detail'
-import { appDetailDataStub, appAuthenticationStub } from '@/sagas/__stubs__/app-detail'
-import { ReduxState } from '@/types/core'
-import appState from '@/reducers/__stubs__/app-state'
+  selectFeaturedApps,
+  selectAppDetail,
+  selectMyApps,
+  selectInstalledApps,
+} from '../app'
 
-describe('app-detail', () => {
+describe('app', () => {
   const mockState = {
     ...appState,
     appDetail: {
@@ -78,6 +83,63 @@ describe('app-detail', () => {
       const input = {} as ReduxState
       const result = selectAppAuthenticationCode(input)
       expect(result).toEqual(undefined)
+    })
+  })
+  describe('selectFeaturedApps', () => {
+    it('should run correctly', () => {
+      const input = {
+        client: {
+          appSummary: {
+            data: {
+              featuredApps: featuredAppsDataStub.data,
+            },
+          },
+        },
+      } as ReduxState
+      const result = selectFeaturedApps(input)
+      expect(result).toEqual(featuredAppsDataStub.data)
+    })
+
+    it('should run correctly and return []', () => {
+      const input = {
+        client: {
+          appSummary: {
+            data: null,
+          },
+        },
+      } as ReduxState
+      const result = selectFeaturedApps(input)
+      expect(result).toEqual([])
+    })
+  })
+
+  describe('selectInstalledApps', () => {
+    it('should selectInstalledApps run correctly and return true', () => {
+      const input = {
+        installedApps: {},
+      } as ReduxState
+      const result = selectInstalledApps(input)
+      expect(result).toEqual({})
+    })
+  })
+
+  describe('selectAppDetail', () => {
+    it('should run correctly and return {}', () => {
+      const input = {
+        appDetail: {},
+      } as ReduxState
+      const result = selectAppDetail(input)
+      expect(result).toEqual(input.appDetail)
+    })
+  })
+
+  describe('selectMyApps', () => {
+    it('should run correctly and return {}', () => {
+      const input = {
+        myApps: {},
+      } as ReduxState
+      const result = selectMyApps(input)
+      expect(result).toEqual(input.myApps)
     })
   })
 })
