+++ packages/marketplace/src/selector/auth.ts
@@ -1,4 +1,5 @@
 import { ReapitConnectSession, LoginIdentity } from '@reapit/connect-session'
+import { COGNITO_GROUP_DEVELOPER_EDITION } from '@/constants/api'
 
 export const selectLoginIdentity = (state: ReapitConnectSession | null): LoginIdentity | undefined => {
   return state?.loginIdentity
@@ -14,4 +15,17 @@ export const selectClientId = (state: ReapitConnectSession | null): string => {
 
 export const selectDeveloperId = (state: ReapitConnectSession | null): string => {
   return state?.loginIdentity.developerId || ''
+}
+
+export const selectLoggedUserEmail = (state: ReapitConnectSession | null): string => {
+  return state?.loginIdentity?.email || ''
+}
+
+export const selectDeveloperEditionId = (state: ReapitConnectSession | null): string | null => {
+  const loginIdentity = selectLoginIdentity(state)
+
+  if (loginIdentity?.groups.includes(COGNITO_GROUP_DEVELOPER_EDITION)) {
+    return state?.loginIdentity?.developerId || ''
+  }
+  return null
 }
