+++ packages/marketplace/src/selector/categories/index.ts
@@ -1,0 +1,1 @@
+export * from './categories'
