+++ packages/marketplace/src/selector/desktop-integration-types/desktop-integration-types.ts
