+++ packages/marketplace/src/selector/desktop-integration-types/index.ts
@@ -1,0 +1,1 @@
+export * from './desktop-integration-types'
