+++ packages/marketplace/src/selector/installations/index.ts
@@ -1,0 +1,1 @@
+export * from './installations'
