+++ packages/marketplace/src/selector/web-components/index.ts
@@ -1,0 +1,1 @@
+export * from './web-components'
